backend/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt>=2.0.0
apscheduler
